Add files via upload
A computer vision based mouse, which can control and command the cursor of a computer or a computerized system using a Web camera. In order to move the cursor on the computer screen the user simply uses an Orange Color within the viewing area of the camera. The video generated by the camera is analyzed using computer vision techniques and the computer moves the cursor according to color movements. The computer vision based mouse also have a feature of Right-click and Left-click.To left-click,I have used a color green and for Right-click,I have used a Yellow color.